Refactor: Rename `layoutCargando` to `loadingContainer` and change type to `FrameLayout`
The variable `layoutCargando` in `TvFragment.kt` and `PeliculasValidasFragment.kt` has been renamed to `loadingContainer` for better clarity.

Additionally, its type has been changed from `LinearLayout` to `FrameLayout`. This change is made because a `FrameLayout` is more appropriate for displaying a single loading view (the ProgressBar and its associated text) that overlays other content. `FrameLayout` is designed to display a single item or stack multiple items on top of each other, which aligns with the loading indicator's behavior.