Refactor: Use CastvHelper for image loading in CardPresenterTV
The image loading logic in `CardPresenterTV`'s `onBindViewHolder` method has been refactored to use the `CastvHelper.loadImage` function.

This change centralizes the image loading mechanism, utilizing Glide with a default placeholder and error drawable. The `ImageView`'s `scaleType` is set to `FIT_XY` and `adjustViewBounds` is set to `false` before loading the image.